Add JSON to library Project.toml
Add JSON to Library project.toml as ONNX profile need to be load with JSON.

+++ ext/JetFlavourTagging/JetConstituentBuilder.jl
@@ -31,7 +31,7 @@ Build the collection of constituents using cluster indices.
 # Returns
 A vector of JetConstituents, each containing the constituents for a specific cluster.
 """
-function build_constituents_cluster(rps::JetConstituents, indices::Vector{Vector{Int}})
+function build_constituents_cluster(rps::JetConstituents, indices::Vector{Vector{Int64}})
     jcs = Vector{JetConstituents}()
     for jet_indices in indices
         jc = rps[jet_indices]
